Moved common class to shared folder
