+++ src/app/gcps-detector.service.ts
@@ -1,11 +1,15 @@
 import { Injectable } from '@angular/core';
-import { CoordsXY } from './common';
+import { CoordsXY } from '../shared/common';
+
+import * as cv from "@techstark/opencv-js";
 
 @Injectable({
     providedIn: 'root'
 })
 export class GcpsDetectorService {
     detect(imgName: string): Promise<CoordsXY> {
+
+        
 
         return Promise.resolve({
             x: 400,
