+++ src/shared/common.ts
